Revise the AEADDecryptor interface.
Finalize() shouldn't depend on the nonce; this is bad for applications.
I modified the interface so that it just takes in the dst to the last
bits of plaintext to. I modified the implementation to buffer at least
gcmTagSize so that the end of the buffer is intepreted as the tag.